submit F.cpp to 'F - Xor Sum 3' (abc141) [C++14 (GCC 5.4.1)]

+++ 2019/0915_ABC141/F.cpp
@@ -283,7 +283,7 @@ int main()
     }
     ++row;
   }
-  int res{0};
+  ll res{0};
   for (auto i = 0; i < row; ++i)
   {
 #if DEBUG == 1
